Bugfix for select with flatten
The select function will, when flatten is enabled, return an array with
all values of the iterable.  This can now be done using
iterable->values().

Moreover, the select function did not properly handle iterables that had
duplicate keys (because of the iterator_to_array implementation).  This
is fixed properly when using the iterable->values() call.