Made reductions\chain behavior stateless
Before reductions\chain is released, I feel this change is needed to
ensure that the reduction is stateless.

This also fixes cases where the data list that is being iterated over
is only a single element long.  This would result in that single
element being returned (as per reduce behavior).  While we expect a
ChainIterator as the reduce result.